+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## XX.XX.XX
+* Added language support for the contents. 
+
 * Mitigated an issue where visibility could have been wrongly assigned if a view was closed while going to background. (Experimental!)
 
 ## 24.7.5